feat(main-api): split to `/me` routes
